+++ src/main/java/roomescape/service/AuthService.java
@@ -1,7 +1,7 @@
 package roomescape.service;
 
 import org.springframework.stereotype.Service;
-import roomescape.auth.TokenProvider;
+import roomescape.auth.token.TokenProvider;
 import roomescape.dto.MemberResponse;
 import roomescape.dto.TokenRequest;
 import roomescape.dto.TokenResponse;
