+++ src/test/java/roomescape/controller/api/ReservationControllerTest.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.auth.TokenProvider;
+import roomescape.auth.token.TokenProvider;
 import roomescape.model.*;
 import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
